Cast to doubles

+++ src/main/java/genepi/imputationserver/steps/InputValidation.java
@@ -233,7 +233,7 @@ public class InputValidation extends WorkflowStep {
 				return false;
 			}
 
-			if (noSnps/chunks > maxChunkSnps) {
+			if ((double)noSnps/(double)chunks > (double)maxChunkSnps) {
 				context.endTask("Your upload data contains " + noSnps + " SNPs in " + chunks + "chunks.\n"
 					+"Input genotypes are expect to come from array genotypes with no more than\n"
 					+ maxChunkSnps + "SNPs expected per chunk.")
